login with email & password added
