+++ src/app/(auth)/login/page.tsx
@@ -30,7 +30,7 @@ import { ArrowLeft } from "lucide-react";
 import { toast } from "sonner";
 
 const formSchema = z.object({
-  username: z.string().min(2).max(20),
+  email: z.string().min(2).max(50).email(),
   password: z.string().min(8).max(50),
 });
 
@@ -55,7 +55,7 @@ const Login = () => {
   const form = useForm<z.infer<typeof formSchema>>({
     resolver: zodResolver(formSchema),
     defaultValues: {
-      username: "",
+      email: "",
       password: "",
     },
   });
@@ -189,12 +189,12 @@ const Login = () => {
                   <div className="flex flex-col gap-2">
                     <FormField
                       control={form.control}
-                      name="username"
+                      name="email"
                       render={({ field }) => (
                         <FormItem>
-                          <FormLabel>Username</FormLabel>
+                          <FormLabel>Email</FormLabel>
                           <FormControl>
-                            <Input placeholder="Username" {...field} />
+                            <Input placeholder="Email Address" {...field} />
                           </FormControl>
                           <FormMessage />
                         </FormItem>
